`trust_remote_code` for hf data components, Default text-gen corpus dataset_type (#728)
## Describe your changes
This PR 
- exposes `trust_remote_code` arguments to huggingface data components
that require user permission to run remote code. For input models with
hf config model_loading_args, the value for `trust_remote_code` is auto
inserted into the data configs without these specified.
- makes the default value for `dataset_type` in
`test_generation_pre_process` component to `corpus`. Previously, the
data config set it to `None` if not given and this results in using
`pair` strategy.
- Casts attention mask to input_id's data type to ensure it has the
expected data type.
- Removes `"dataset_type":"corpus"` from the examples since it is the
default and the most common type.
- Corrects the dataset config in falcon example. Dataset is also changed
to `timdettmers/openassistant-guanaco` since the previous dataset is
massive (>1TB) and we only need it for latency measurement.
 
## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [x] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link

+++ olive/data/component/text_generation.py
@@ -542,7 +542,9 @@ def append_text_gen_input_ids(
     # create attention_mask
     if use_attention_mask:
         attention_mask = (
-            torch.ones_like(input_ids) if tokenizer.pad_token_id is None else input_ids.ne(tokenizer.pad_token_id)
+            torch.ones_like(input_ids)
+            if tokenizer.pad_token_id is None
+            else input_ids.ne(tokenizer.pad_token_id).to(input_ids.dtype)  # is boolean otherwise
         )
         inputs["attention_mask"] = attention_mask
